prepare dependency-injection
