+++ Cargo.toml
@@ -57,6 +57,7 @@ url = { version = "2.5", optional = true }
 zombienet-sdk = { git = "https://github.com/paritytech/zombienet-sdk", optional = true }
 zombienet-support = { git = "https://github.com/paritytech/zombienet-sdk", optional = true }
 uuid = { version = "1.7.0", features = ["v4"] }
+toml = "0.8.10"
 
 
 [features]
